Verision 1.1
Added Masternode Genkey command. 
Now h2o.conf only gets updated with the new genkey
All dat files are removed. 
Some new checks for systemd and other processes.

+++ h2o_upgrade_mn.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-usesystemd=0
+BINPATH=$(which h2od | sed "s/h2od//g")
+USESYSTEMD=$(systemctl list-unit-files|grep h2od|wc -l)
+
 ##---Check if root use may need sudo in some spots
 if [ $(id -u) -ne 0 ]
 then
@@ -9,36 +11,48 @@ else
   SUDO=''
 fi
 
-##---Try to stop h2o service
-h2o-cli stop
-echo "Trying to stop h2o service..."
-sleep 6
-
-#might be using systemd
-if pgrep -x "h2od" > /dev/null
-then
-    $SUDO systemctl stop h2o
-    echo "Trying to stop h2o service... again"
-    sleep 6
-    if ! pgrep -x "h2od" > /dev/null
-    then
-        usesystemd=1
-    fi
+if sudo test -d "/root/.h2ocore"; then
+	DATPATH="/root/.h2ocore"
+else 
+	if test -d "$HOME/.h2ocore"; then
+		DATPATH="$HOME/.h2ocore"
+	else
+		echo -e "\nUnable to locate H2O data file path."
+		exit 1
+	fi
 fi
 
 if pgrep -x "h2od" > /dev/null
 then
-    printf "\nCannot stop masternode."
-    printf "\nPlease shutdown H2O masternode before installing."
-    printf "\nUse the command h2o-cli stop \n\n"
-    exit 0
+	MNPKEY=$($BINPATH/h2o-cli masternode genkey)
+	
+	echo "Trying to stop h2o service..."
+
+	if  [ "$USESYSTEMD" -eq "1" ];
+	then
+		$SUDO systemctl stop h2o
+	else
+		$BINPATH/h2o-cli stop
+	fi
+	
+	sleep 6
+
+	if pgrep -x "h2od" > /dev/null
+	then
+		echo -e "\nCannot stop masternode."
+		echo -e "\nPlease shutdown H2O masternode before installing."
+		echo -e "\nUse the command h2o-cli stop \n\n"
+		exit 1
+	fi
+else
+	MNPKEY=$(egrep "^masternodeprivkey=" $DATPATH/h2o.conf |awk -F '=' '{print $2}')
 fi
 
 ##---Download new H2O Wallet and uncompress
 echo "Downloading upgraded h2o wallet..."
 wget https://github.com/h2ocore/h2o/releases/download/v0.12.1.7/Linux64-H2O-cli-01217.tgz
 tar -zxf Linux64-H2O-cli-01217.tgz
-rm Linux64-H2O-cli-01217.tgz
+rm -f Linux64-H2O-cli-01217.tgz
 
 ##---Check that files exist
 f1="h2od"
@@ -61,39 +75,60 @@ if [ ! -f $f3 ]; then
     exit 0
 fi
 
-##---Check directory for current h2o binaries and move to that directory
-mvpath=$(which h2od | sed "s/h2od//g")
+echo "Moving {$f1} {$f2} {$f3} to $BINPATH..."
 
-echo "Moving {$f1} {$f2} {$f3} to $mvpath..."
+$SUDO mv -f $f1 $f2 $f3 $BINPATH
 
-$SUDO mv $f1 $f2 $f3 $mvpath
+$SUDO rm -f $DATPATH/banlist.dat
+$SUDO rm -f $DATPATH/fee_estimates.dat
+$SUDO rm -f $DATPATH/governance.dat
+$SUDO rm -f $DATPATH/mncache.dat
+$SUDO rm -f $DATPATH/mnpayments.dat
+$SUDO rm -f $DATPATH/netfulfilled.dat
+$SUDO rm -f $DATPATH/peers.dat
+
+$SUDO cp $DATPATH/h2o.conf $DATPATH/h2o.conf.$(date +%Y%m%d%H%M%S).saved 
+$SUDO egrep -v "^masternodeprivkey=" $DATPATH/h2o.conf >/tmp/h2o.conf.new
+echo -e "\nmasternodeprivkey=$MNPKEY\n" >>/tmp/h2o.conf.new
+$SUDO mv -f /tmp/h2o.conf.new $DATPATH/h2o.conf
+
+echo -e "************************************************************************"
+echo -e "***"
+echo -e "***   A new masternode private key has been generated"
+echo -e "***   Please update your 'masternode.conf' file with this new key:"
+echo -e "***"
+echo -e "***                $MNPKEY"
+echo -e "***"
+echo -e "************************************************************************\n\n"
 
 ##---Restart H2O Masternode
 echo "Trying to restart masternode..."
-if  [ "$usesystemd" -eq "1" ];
+if  [ "$USESYSTEMD" -eq "1" ];
 then
     $SUDO systemctl start h2o
 else
-    h2od
+    $BINPATH/h2od
 fi
 
-#sleep 10
+sleep 3
 
 if ! pgrep -x "h2od" > /dev/null
 then
     echo "Could not restart H2O masternode :'("
-    exit 0
+    exit 1
 fi
 
 echo "Checking version..."
-sleep 6
+sleep 2
+
 ##---Check for success for failure of upgrade
 if  h2o-cli getinfo | grep -m 1 '"protocolversion": 70209,'
 then
-        echo "H2O Mastenode has been successfully uppdated"
+        echo "H2O Mastenode has been successfully updated"
 else
         echo "H2O masternode Install failed :'( "
 fi
-printf "\n\nUse h2o-cli mnsync status\n"
-echo "and h2o-cli masternode status"
-echo "to verify masternode is active"
+
+echo -e "\n\nUse\n\th2o-cli mnsync status\n"
+echo -e "and\n\th2o-cli masternode status\n"
+echo -e "to verify if masternode is active"
